Add behavior to go to next stage

+++ backend/bracket/routes/courts.py
@@ -1,5 +1,6 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from heliclockter import datetime_utc
+from starlette import status
 
 from bracket.database import database
 from bracket.models.db.court import Court, CourtBody, CourtToInsert
@@ -8,6 +9,7 @@ from bracket.routes.auth import user_authenticated_for_tournament
 from bracket.routes.models import CourtsResponse, SingleCourtResponse, SuccessResponse
 from bracket.schema import courts
 from bracket.sql.courts import get_all_courts_in_tournament, update_court
+from bracket.sql.stages import get_stages_with_rounds_and_matches
 from bracket.utils.db import fetch_one_parsed
 from bracket.utils.types import assert_some
 
@@ -51,6 +53,20 @@ async def update_court_by_id(
 async def delete_court(
     tournament_id: int, court_id: int, _: UserPublic = Depends(user_authenticated_for_tournament)
 ) -> SuccessResponse:
+    stages = await get_stages_with_rounds_and_matches(tournament_id, no_draft_rounds=False)
+    used_in_matches_count = 0
+    for stage in stages:
+        for round_ in stage.rounds:
+            for match in round_.matches:
+                if match.court_id == court_id:
+                    used_in_matches_count += 1
+
+    if used_in_matches_count > 0:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"Could not delete court since it's used by {used_in_matches_count} matches",
+        )
+
     await database.execute(
         query=courts.delete().where(
             courts.c.id == court_id and courts.c.tournament_id == tournament_id
